+++ src/browser/js/app/managers/state.js
@@ -85,7 +85,7 @@ var StateManager = class StateManager {
         var prompt = DOM.create('<input type="file" accept=".state"/>')
 
         prompt.click()
-        prompt.addEventListener('change',function(e){
+        prompt.addEventListener('change',(e)=>{
 
             var reader = new FileReader(),
                 loader = loading('Uploading file...')
@@ -98,7 +98,7 @@ var StateManager = class StateManager {
             reader.onload = (e)=>{
                 loader.close()
                 this.set(JSON.parse(e.target.result),true)
-                this.state = preset
+                this.state = e.target.result
 
             }