+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - bug fixes
   - error when resizing a widget and its clone(s) at the same time
   - error with non string css property
-
+  - session list not reappearing when failing to load
+  
 ## 0.27.0
 
 - bug fixes