+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - error when resizing a widget and its clone(s) at the same time
   - error with non string css property
   - session list not reappearing when failing to load
+  - broken state file loading
 - misc
   - added a loading indicator when uploading a session file or a state file